Testing deployment

+++ src/app/components/edit-statements/edit-statements.component.ts
@@ -13,6 +13,7 @@ import { SurveyService } from '../../survey.service';
 export class EditStatementsComponent implements OnInit {
 
   @Input() statements: any[];
+  @Input() disabled: boolean;
   @Output() status = new EventEmitter<boolean>();
 
   angForm: FormGroup;
@@ -31,22 +32,38 @@ export class EditStatementsComponent implements OnInit {
     }
 
   addStatement(statement: string) {
-    this.route.params.subscribe(params => {
-      // console.log(statement);
-      this.surveyservice.addStatement(params['id'], statement).subscribe(res => {
-        this.status.emit(true);
+    if (!this.disabled) {
+      this.route.params.subscribe(params => {
+        // console.log(statement);
+        this.surveyservice.addStatement(params['id'], statement).subscribe(res => {
+          this.status.emit(true);
+        });
       });
-    });
+    } else {
+      try {
+        throw new Error('Attempted to update a published server');
+      } catch (e) {
+        alert(e.name + ': ' + e.message);
+      }
+    }
   }
 
   deleteStatement(statement_id) {
-    this.route.params.subscribe(params => {
-      if (window.confirm('Are you sure you wish to delete this statement?')) {
-        this.surveyservice.deleteStatement(params['id'], statement_id).subscribe(res => {
-          this.status.emit(true);
-        });
+    if (!this.disabled) {
+      this.route.params.subscribe(params => {
+        if (window.confirm('Are you sure you wish to delete this statement?')) {
+          this.surveyservice.deleteStatement(params['id'], statement_id).subscribe(res => {
+            this.status.emit(true);
+          });
+        }
+      });
+    } else {
+      try {
+        throw new Error('Attempted to update a published server');
+      } catch (e) {
+        alert(e.name + ': ' + e.message);
       }
-    });
+    }
   }
 
   ngOnInit() {
